D3DFMIQ-1895 Merged revision(s) 48182 from branches/FeatureBranches/NGHS_BO_DataAccessLayerFM: Fixed Sonarqube smells. ........
Former-commit-id: 26b3fdaa43b7fd3cb082f421d61ed13f1f347866
Former-commit-id: 68ed67644196502460233ef026002f459ed29e05

+++ src/DeltaShell.Plugins.FMSuite.FlowFM/WaterFlowFMModelCoordinateConversion.cs
@@ -84,8 +84,8 @@ namespace DeltaShell.Plugins.FMSuite.FlowFM
                 double[] newCellCenter =
                     transformation.MathTransform.Transform(new[]
                     {
-                        (double) cell.CenterX,
-                        (double) cell.CenterY
+                        cell.CenterX,
+                        cell.CenterY
                     });
                 cell.CenterX = (float) newCellCenter[0];
                 cell.CenterY = (float) newCellCenter[1];
